requirements.txt
django
djangorestframework
environs
secrets
requests
psycopg2